notebooks/requirements.txt
umap-learn>=0.5
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
plotly>=5.0
scipy>=1.11
